requirements.txt
Flask==3.1.3
python-docx==1.2.0
PyMuPDF==1.27.2.2
python-dotenv==1.2.2
pywin32==311

# Google Drive + OAuth
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1

# Importação via URL
requests==2.33.1
beautifulsoup4==4.14.3

# Frontmatter .md
python-frontmatter==1.1.0
